+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][]
 
+- feat(grid-carousel): added grid-carousel component
+
 ## [1.39.0][] - 2023-06-26
 
 - feat(api-template-vue): added simple state mechanic for interactive components