+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][]
 
+- feat(grid-carousel): added grid-carousel component
+- fix(deps): updated lerna to 7.x
+
 ## [1.39.0][] - 2023-06-26
 
 - feat(api-template-vue): added simple state mechanic for interactive components
@@ -1416,7 +1419,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **content-carousel**: Initial version of content carousel, a carousel based portlet metadata display system.
 
 [unreleased]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.39.0...HEAD
-[1.37.0]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.38.1...v1.39.0
+[1.39.0]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.38.1...v1.39.0
+[1.38.1]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.38.0...v1.38.1
+[1.38.0]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.37.0...v1.38.0
 [1.37.0]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.36.0...v1.37.0
 [1.36.0]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.35.0...v1.36.0
 [1.35.0]: https://github.com/uPortal-contrib/uPortal-web-components/compare/v1.34.3...v1.35.0